[web] Refactor: group asset-ratio-section related components
