+++ packages/bento-web/src/dashboard/sections/AssetRatioSection/index.ts
@@ -1,0 +1,1 @@
+export { AssetRatioSection } from './AssetRatioSection';
